cpu code should be in double

+++ src/cpu/aabb.cpp
@@ -10,13 +10,13 @@ void addEdges
     vector<Aabb>& boxes
 )
 {
-    for (unsigned long i = 0; i < edges.rows(); i++) {
-        Eigen::MatrixXf edge_vertex0_t0 = vertices_t0.cast<float>().row(edges(i, 0));
-        Eigen::MatrixXf edge_vertex1_t0 = vertices_t0.cast<float>().row(edges(i, 1));
-        Eigen::MatrixXf edge_vertex0_t1 = vertices_t1.cast<float>().row(edges(i, 0));
-        Eigen::MatrixXf edge_vertex1_t1 = vertices_t1.cast<float>().row(edges(i, 1));
+      for (unsigned long i = 0; i < edges.rows(); i++) {
+        Eigen::MatrixXd edge_vertex0_t0 = vertices_t0.row(edges(i, 0));
+        Eigen::MatrixXd edge_vertex1_t0 = vertices_t0.row(edges(i, 1));
+        Eigen::MatrixXd edge_vertex0_t1 = vertices_t1.row(edges(i, 0));
+        Eigen::MatrixXd edge_vertex1_t1 = vertices_t1.row(edges(i, 1));
 
-        Eigen::MatrixXf points(4, edge_vertex0_t0.size());
+        Eigen::MatrixXd points(4, edge_vertex0_t0.size());
         points.row(0) = edge_vertex0_t0;
         points.row(1) = edge_vertex1_t0;
         points.row(2) = edge_vertex0_t1;
@@ -24,8 +24,8 @@ void addEdges
 
         int vertexIds[3] = {edges(i,0), edges(i,1), -edges(i,0)-1};
 
-        Eigen::MatrixXf lower_bound = points.colwise().minCoeff();
-        Eigen::MatrixXf upper_bound = points.colwise().maxCoeff();
+        Eigen::MatrixXd lower_bound = points.colwise().minCoeff();
+        Eigen::MatrixXd upper_bound = points.colwise().maxCoeff();
         boxes.emplace_back(boxes.size(), i, Simplex::EDGE, vertexIds, lower_bound.array().data(), upper_bound.array().data());
     }
 }
@@ -37,18 +37,18 @@ void addVertices
     vector<Aabb>& boxes
 )
 {
-    for (unsigned long i = 0; i < vertices_t0.rows(); i++) {
-        Eigen::MatrixXf vertex_t0 = vertices_t0.cast<float>().row(i);
-        Eigen::MatrixXf vertex_t1 = vertices_t1.cast<float>().row(i);
+   for (unsigned long i = 0; i < vertices_t0.rows(); i++) {
+        Eigen::MatrixXd vertex_t0 = vertices_t0.row(i);
+        Eigen::MatrixXd vertex_t1 = vertices_t1.row(i);
 
-        Eigen::MatrixXf points(2, vertex_t0.size());
+        Eigen::MatrixXd points(2, vertex_t0.size());
         points.row(0) = vertex_t0;
         points.row(1) = vertex_t1;
 
         int vertexIds[3] = {i, -i-1, -i-1};
 
-        Eigen::MatrixXf lower_bound = points.colwise().minCoeff();
-        Eigen::MatrixXf upper_bound = points.colwise().maxCoeff();
+        Eigen::MatrixXd lower_bound = points.colwise().minCoeff();
+        Eigen::MatrixXd upper_bound = points.colwise().maxCoeff();
         boxes.emplace_back(boxes.size(), i, Simplex::VERTEX, vertexIds, lower_bound.array().data(), upper_bound.array().data());
     }
 }
@@ -62,14 +62,14 @@ void addFaces
 )
 {
     for (unsigned long i = 0; i < faces.rows(); i++) {
-        Eigen::MatrixXf face_vertex0_t0 = vertices_t0.cast<float>().row(faces(i, 0));
-        Eigen::MatrixXf face_vertex1_t0 = vertices_t0.cast<float>().row(faces(i, 1));
-        Eigen::MatrixXf face_vertex2_t0 = vertices_t0.cast<float>().row(faces(i, 2)); 
-        Eigen::MatrixXf face_vertex0_t1 = vertices_t1.cast<float>().row(faces(i, 0));
-        Eigen::MatrixXf face_vertex1_t1 = vertices_t1.cast<float>().row(faces(i, 1)); 
-        Eigen::MatrixXf face_vertex2_t1 = vertices_t1.cast<float>().row(faces(i, 2));
+        Eigen::MatrixXd face_vertex0_t0 = vertices_t0.row(faces(i, 0));
+        Eigen::MatrixXd face_vertex1_t0 = vertices_t0.row(faces(i, 1));
+        Eigen::MatrixXd face_vertex2_t0 = vertices_t0.row(faces(i, 2)); 
+        Eigen::MatrixXd face_vertex0_t1 = vertices_t1.row(faces(i, 0));
+        Eigen::MatrixXd face_vertex1_t1 = vertices_t1.row(faces(i, 1)); 
+        Eigen::MatrixXd face_vertex2_t1 = vertices_t1.row(faces(i, 2));
 
-        Eigen::MatrixXf points(6, face_vertex0_t0.size());
+        Eigen::MatrixXd points(6, face_vertex0_t0.size());
         points.row(0) = face_vertex0_t0;
         points.row(1) = face_vertex1_t0;
         points.row(2) = face_vertex2_t0;
@@ -79,8 +79,8 @@ void addFaces
 
         int vertexIds[3] = {faces(i,0), faces(i,1), faces(i,2)};
 
-        Eigen::MatrixXf lower_bound = points.colwise().minCoeff();
-        Eigen::MatrixXf upper_bound = points.colwise().maxCoeff();
+        Eigen::MatrixXd lower_bound = points.colwise().minCoeff();
+        Eigen::MatrixXd upper_bound = points.colwise().maxCoeff();
         boxes.emplace_back(boxes.size(), i, Simplex::FACE, vertexIds, lower_bound.array().data(), upper_bound.array().data());
     }
 };
